create worldAI
